requirements.txt
pandas>=2.0.0
pytest>=8.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
fairlearn>=0.9.0
imbalanced-learn>=0.11.0
xgboost>=2.0.0
matplotlib>=3.7.0
kagglehub>=0.2.0
google-genai>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
websockets>=11.0
# Optional: for PDF from Markdown when pdflatex is not installed
markdown>=3.4.0
fpdf2>=2.8.0
# Research pipeline: alphaXiv Assistant V2 + arXiv/Semantic Scholar fallback
httpx>=0.25.0
python-dotenv>=1.0.0
arxiv>=2.0.0
semanticscholar>=0.8.0
pypdf>=4.0.0

# Security pins (CVE fixes) — keep updated
filelock>=3.20.3
requests>=2.32.4
urllib3>=2.6.3
